lost_found/requirements.txt
Flask==1.0
Flask-SQLAlchemy==2.3.2
Flask-Testing==0.6.2
gunicorn==19.7.1
coverage==4.4.2
flake8===3.5.0
mysqlclient
flask-migrate==2.1.1
Flask-JSON
pyjwt==1.5.3
flask-bcrypt==0.7.1
itsdangerous
flask_mail